requirements.txt
streamlit
tensorflow==2.16.2
numpy
pillow
requests
streamlit-lottie
